Address review: honor fernet-aes128 feature, fix advertised MTU, module size
Two issues from automated review, plus one we found independently while
continuing to test live against a real destination:

1. LinkMode::DEFAULT was a fixed constant (Aes256Cbc) regardless of which
   AES key size this build was actually compiled with. crypt::fernet's
   AesAlgo is a compile-time type alias gated by the fernet-aes128
   feature, so a build compiled with that feature enabled would advertise
   (and, on repeated timeouts, fall back to) a mode it can never actually
   encrypt/decrypt with — a peer could accept and prove such a Link, which
   would then activate but be unable to exchange any data. DEFAULT is now
   #[cfg]-selected to match the compiled-in cipher, and the fallback
   mechanism is removed entirely: with only one cipher ever compiled into
   a given build, there is no other mode to usefully fall back to, so
   every request now consistently signals the one mode this build can
   actually use.

2. Moved LinkMode into its own module_prelude-adjacent file
   (link_parts/link_mode.rs) — the previous version pushed new.rs to 516
   lines against this repo's 500-line module budget
   (tools/scripts/check-module-size.sh).

3. Continued live testing against the same real destination from the
   linked issue turned up a second, independent problem with the original
   fix: advertising a large explicit MTU (8192, copied from a wire capture
   of another client's request) reliably let the Link itself activate but
   then broke the single-packet Request/Response exchange on top of it.
   Re-testing with a small, explicit MTU instead succeeded cleanly and
   repeatedly. The advertised value is now RETICULUM_COMPAT_MTU — this
   crate's own existing conservative ceiling, already used to clamp an
   *incoming* signalled value, reused here rather than inventing a new
   constant.

Tests updated to match (no more fallback-behavior test, since there's
nothing to fall back to; a new test confirms the mode never changes
across repeated requests). Full suite passes under both the default and
fernet-aes128 feature configurations; module-size check, fmt, and clippy
all clean under both as well.
