+++ crates/libs/rns-transport/src/destination/link_parts/link_status_predicates_make_lifecycl_sections/core_tests.rs
@@ -88,45 +88,28 @@
     }
 
     #[test]
-    fn outbound_link_request_defaults_to_aes_256_cbc_and_the_advertised_mtu() {
+    fn outbound_link_request_signals_the_compiled_in_mode_and_a_conservative_mtu() {
         let mut link = test_link();
         let request = link.request();
 
         assert_eq!(request.data.len(), PUBLIC_KEY_LENGTH * 2 + LINK_MTU_SIZE);
         let (mode_bits, mtu) = decode_signalling(&request);
         assert_eq!(mode_bits, LinkMode::DEFAULT.mode_bits());
-        assert_eq!(LinkMode::DEFAULT, LinkMode::Aes256Cbc);
-        assert_eq!(mtu, LINK_ADVERTISED_MTU);
+        assert_eq!(mtu, RETICULUM_COMPAT_MTU);
     }
 
     #[test]
-    fn outbound_link_request_falls_back_to_the_other_mode_after_repeated_timeouts() {
-        let mut link = test_link();
-
+    fn outbound_link_request_never_changes_mode_across_repeated_attempts() {
         // The watchdog's own repeat-request cadence calls `request()`
-        // again on every retry — simulate exactly that, with no
-        // successful Proof in between (LinkStatus stays Pending). The Nth
-        // call is the one that increments the counter PAST the threshold
-        // and flips within that same call, so exactly
-        // `MODE_FALLBACK_ATTEMPTS` calls stay on the default mode before
-        // the very next one flips.
-        for _ in 0..MODE_FALLBACK_ATTEMPTS {
+        // again on every retry — with only one cipher ever compiled into
+        // a given build (see `LinkMode`'s own doc comment), there is no
+        // other mode to fall back to, so every repeat must keep signalling
+        // the same, one mode this build can actually use.
+        let mut link = test_link();
+        for _ in 0..5 {
             let (mode_bits, _) = decode_signalling(&link.request());
-            assert_eq!(mode_bits, LinkMode::Aes256Cbc.mode_bits(), "should stay on the default mode until the threshold is exceeded");
+            assert_eq!(mode_bits, LinkMode::DEFAULT.mode_bits());
         }
-        let (mode_bits, _) = decode_signalling(&link.request());
-        assert_eq!(mode_bits, LinkMode::Aes128Cbc.mode_bits(), "should fall back once MODE_FALLBACK_ATTEMPTS is exceeded");
-
-        // And it round-trips back, the same way, if the fallback mode also
-        // doesn't get a Proof in time — there being only two known modes,
-        // repeated exhaustion just alternates between them forever, which
-        // is exactly what "try both, keep trying both" should do.
-        for _ in 0..MODE_FALLBACK_ATTEMPTS {
-            let (mode_bits, _) = decode_signalling(&link.request());
-            assert_eq!(mode_bits, LinkMode::Aes128Cbc.mode_bits());
-        }
-        let (mode_bits, _) = decode_signalling(&link.request());
-        assert_eq!(mode_bits, LinkMode::Aes256Cbc.mode_bits(), "should fall back to the original mode again after another full round");
     }
 
     #[test]
